Add files via upload
The problem of fake companies on the LinkedIn platform posting job listings for
personal information collection and identity theft, as well as a data breach where
the fake profile companies obtained user information from the resumes. To combat
these issues, we have created LinkedGuard, an approach to safeguarding LinkedIn
privacy. Our project is divided into two parts, the first part of the project is to
find out the legitimacy of the company and predict the risk label using Machine
Learning. The second part of the project is the scrapper part, wherein we scrape
the resume based on risk label prediction for each company.